refactor(template): Snippet 도메인을 Template 하위로 이동
