+++ backend/src/main/java/codezap/template/repository/SnippetRepository.java
@@ -1,10 +1,10 @@
-package codezap.snippet.repository;
+package codezap.template.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import codezap.snippet.domain.Snippet;
+import codezap.template.domain.Snippet;
 import codezap.template.domain.Template;
 
 public interface SnippetRepository extends JpaRepository<Snippet, Long> {
